BA: implement basic preemptive defense
- switch defendhandler to use runfromattack behaviour, and speed it up
by making it use runfromattack's underfire rather than sending a danger
signal. make danger signal only for buildings (i.e., turtles)
- make defendhandler and defendbehaviour use turtles as defendees. use
the intersection with the edge of turtle to determine where defenders
should be placed.
-  add a registry of turtlesByUnitID so that Danger() will add the
associated turtle as a defendee instead of the building itself (which is
a terrible behaviour)